Avoid parsing loc.gov enchancment URLs as Items
Even though http://www.loc.gov/marc/bibliographic/bd856.html shows an example
where an 856 field with ind2=1 links to a table of contents and is considered a
"Version of resource", no reasonable human would consider the table of contents a bf:Item
or an electronic version of the resource at hand.

Thus, avoid generating a bf:Item for any loc.gov enhancement URLs, and treat is
as bf:SupplementaryContent instead.

Signed-off-by: Dan Scott <dscott@laurentian.ca>